requirements.txt
Beaker==1.7.0
Chameleon==2.22
Mako==1.0.2
MarkupSafe==0.23
PasteDeploy==1.5.2
PyJWT==1.4.0
Pygments==2.0.2
WebOb==1.4.1
anykeystore==0.2
argparse==1.2.1
elasticsearch==1.7.0
future==0.15.2
gevent==1.1b4
greenlet==0.4.9
gunicorn==19.3.0
oauthlib==1.0.3
py-bcrypt==0.4
pycrypto==2.6.1
pymongo==2.7.2
pyramid==1.5.7
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.1
pyramid-mako==1.0.2
python-openid==2.2.5
repoze.lru==0.6
requests==2.7.0
requests-oauthlib==0.5.0
translationstring==1.3
urllib3==1.12
velruse==1.1.1
venusian==1.0
waitress==0.8.10
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.2
